Docker installation script was improved.
Former-commit-id: b0fde0afd8b0bc97e3ba695af15b0f9b12ca4ce5 [formerly d9331a05709d59c58600c31fad303fae0d628bf4] [formerly 400e87db6a80c49a9814a711cc38d9c9a3726cf5 [formerly e572ee4faec5028a70c851fc98f9e7681c308763 [formerly e572ee4faec5028a70c851fc98f9e7681c308763 [formerly e572ee4faec5028a70c851fc98f9e7681c308763 [formerly f75a1b409e4572a28f83b06ab0380cdfad6aa38c]]]]]
Former-commit-id: 7267875f893c8d8e3f8060241dd202c6b99f3c60 [formerly f361ddc93cd5c8babd61fd5d4269b5046421680c]
Former-commit-id: 40a7cd666fe6f2b7a3a986a844ae8e1fd3e06e9e
Former-commit-id: 28b6d6322afd36892b16ac3ce32daea9ad2c34e8
Former-commit-id: 492643cf1ed526bdc35f34f2dadbf08f3a700ea3
Former-commit-id: 126ab4f012f0f9c0b8920a5422aca98b2faa14e2

+++ scripts/docker.sh
@@ -28,5 +28,7 @@ apt-get install docker-engine
 service docker start
 
 echo "[docker] Installing docker-compose"
-curl -L "https://github.com/docker/compose/releases/download/1.8.1/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
-chmod +x /usr/local/bin/docker-compose
+if [ ! -e /usr/local/bin/docker-compose ]; then
+	curl -L "https://github.com/docker/compose/releases/download/1.8.1/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
+	chmod +x /usr/local/bin/docker-compose
+fi
